Added some new code to implement a permission scheme.

+++ app/Http/Controllers/CurrencyController.php
@@ -1,5 +1,6 @@
 <?php namespace FireflyIII\Http\Controllers;
 
+use Auth;
 use Cache;
 use FireflyIII\Http\Requests\CurrencyFormRequest;
 use FireflyIII\Models\TransactionCurrency;
@@ -107,8 +108,9 @@ class CurrencyController extends Controller
         }
 
         Session::flash('success', 'Currency "' . e($currency->name) . '" deleted');
-
-        $currency->delete();
+        if (Auth::user()->hasRole('owner')) {
+            $currency->delete();
+        }
 
         return Redirect::to(Session::get('currency.delete.url'));
     }
@@ -146,6 +148,12 @@ class CurrencyController extends Controller
         $currencies      = $repository->get();
         $defaultCurrency = $repository->getCurrencyByPreference(Preferences::get('currencyPreference', 'EUR'));
 
+
+        if (!Auth::user()->hasRole('owner')) {
+            Session::flash('warning', 'Please ask ' . env('SITE_OWNER') . ' to add, remove or edit currencies.');
+        }
+
+
         return view('currency.index', compact('currencies', 'defaultCurrency'));
     }
 
@@ -158,11 +166,11 @@ class CurrencyController extends Controller
      */
     public function store(CurrencyFormRequest $request, CurrencyRepositoryInterface $repository)
     {
-        $data     = $request->getCurrencyData();
-        $currency = $repository->store($data);
-
-
-        Session::flash('success', 'Currency "' . $currency->name . '" created');
+        $data = $request->getCurrencyData();
+        if (Auth::user()->hasRole('owner')) {
+            $currency = $repository->store($data);
+            Session::flash('success', 'Currency "' . $currency->name . '" created');
+        }
 
         if (intval(Input::get('create_another')) === 1) {
             Session::put('currency.create.fromStore', true);
@@ -185,9 +193,10 @@ class CurrencyController extends Controller
      */
     public function update(CurrencyFormRequest $request, CurrencyRepositoryInterface $repository, TransactionCurrency $currency)
     {
-        $data     = $request->getCurrencyData();
-        $currency = $repository->update($currency, $data);
-
+        $data = $request->getCurrencyData();
+        if (Auth::user()->hasRole('owner')) {
+            $currency = $repository->update($currency, $data);
+        }
         Session::flash('success', 'Currency "' . e($currency->name) . '" updated.');
